Bug 532464 - Opening a new window destroys/blocks tab dragging in first window
Fixed search order for the control search. The active shell should be
checked first, otherwise the findControl() could find "visible" widgets
from shells which are overlapped with the active one.

Reduced visibility of findControl() methods to avoid cases clients
search for controls in a wrong order. The
org.eclipse.e4.ui.workbench.addons.dndaddon package is not API, so
should be OK to do so.

Change-Id: Ia7e4f48f76cc043a6e905e892745ae13e141c18b
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>