Bug 485167 Bring UI Test Suite launch target in line with Tycho configuration
When the UI Test Suite is launched from Tycho, the test plug-in
dependencies are used for resolving the plug-ins that are included in
the launch target. When the Eclipse launch target is used, all plug-ins
in the current workspace are used. As a result of the different test
environments some test cases fail when run with the Eclipse launch
target: ProgressViewTests, ObjectContributionTest, among others.

This commit changes the Eclipse launch target to be in line with the
Tycho behavior. A new feature is introduced to define the the test
runtime platform configuration. The launch target uses only this feature
to determine the runtime platform.

Change-Id: I5d18778ee2e40bb55b16378a6a52df70a8efccf2
Signed-off-by: Rolf Theunissen <rolf.theunissen@gmail.com>