Bug 539661 - Restores the Search view background color
25715a37d79ac0b1a79852355f72a31e4095fda9 set the background color for
Composites in CTabFolder to inherit which breaks for example the Search
Dialog (Ctrl+4) initial background color.

This change removes the setting for the toolbar to fix the SearchDialog
and similar dialogs. A followup commit is needed to target the Composite
for the toobar. Most likely this will require to set a CSS class on this
Composite to allow to target it.


Change-Id: If6e788c1f1ea388d829cbc3549a1e8d14d5779c8
Signed-off-by: Lars Vogel <Lars.Vogel@vogella.com>