Bug 546187 - [Quick Access] Delay provider instantiation and loading
Loading QuickAccessProviders was happening when building the
SearchField, during workspace initialization. This was causing soem
cycles spent at initialization, and -more critical- loading of
extensions and related plugins at workbench startup.
The loading of providers and related content and widgets in SearchField
is now delayed to the 1st time SearchField receives focus.

Change-Id: Ie71fbb6f8a57b56b8670e6e79ea4177c6f37ee03
Signed-off-by: Mickael Istria <mistria@redhat.com>