Bug 546710 - [console] Synchronize "remove all" action enable update
This should fix the following race condition for fast terminating
processes:
When the action is constructed it will update its enablement. The
terminated notification, which also updates enablement, can come from a
different thread.
It may happen that the initial enablement update is interrupted after
calculating the enablement state but before setting it. If then the
process terminates the action is enabled due to terminated notification
but the later resumed initial update will set it to disabled again.

Change-Id: I32eafe4676b3fbd57a997ef75c96965bb5159f2f
Signed-off-by: Paul Pazderski <paul-eclipse@ppazderski.de>