Bug 546828 - Editors based on FileDocumentProvider / TextFileDocumentProvider lock UI on attempt to edit while build is running
The change modifies FileDocumentProvider.getValidateStateRule() and
DocumentProviderOperation created in
TextFileDocumentProvider.validateState() to
return non-null rule (referenced IFile) for IFileEditorInput's if the
IResourceRuleFactory.validateEditRule() has no rule for us.

This prevents UI freeze on attempt to modify a file while build is
running. Instead of a freeze, we get a progress dialog with a
possibility to cancel the (blocking) attempt to modify the file.

Change-Id: I23a6ddab2ca8f31b5ba134301356c34250f5087c
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>