Bug 547001 - quickAccess extension only for active contributor bundle
The general idea is to create a proxy delegating to extensions and only
loading the code when the contributing bundle is active.
The QuickAccessProvider and QuickAccessElement class were too stateful
to handle that, so the extension and classes were refactored to be more
stateless and with a reduced scope. Most non-dynamic features of the
extensions now is to be contributed statically in the extension via
plugin.xml, so the class doesn't have to be loaded to get its
"metadata", and bundle can remain inactive.

Change-Id: Ia50a705358112ecfa94057bbecd1b89cc2c9b359
Signed-off-by: Mickael Istria <mistria@redhat.com>