Bug 485896 - PartRenderingEngine#fixZOrder triggers synchronous layout
Synchronous layout calls should be avoided. Especially during startup,
as this only delays the renderering of the UI without any benefit for
the user as the UI is not yet available.

To test: close and open views and editor and check that the new element
is rendered correctly. This method is also frequently called during
startup for Toolbars, so ensure that the toolbars are renderered
correctly after starting a runtime Eclipse.

Change-Id: Ide92f98d7296eb3676332bed00e42048693703d8
Signed-off-by: Lars Vogel <Lars.Vogel@vogella.com>