Bug 547157 - Fix Show View dialog element ordering
The ViewComparator was incorrectly comparing different strings to
the different sides of the comparison when trying to sort 'General'
first. This resulted in an effectively random sorting based on which
elements the sort algorithm decided to compare (and in which order).

If the 'Other' category is intended to be placed last (which would seem
reasonable), this would be easy to achieve with the same logic, but
I could see nothing suggesting this is the intention other than the one
erroneous line of code.

Change-Id: I632cf091ce5aa48c1942c33913631046a4451df7
Signed-off-by: Jonathan Simmonds <jonathan.simmonds@arm.com>
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>