Bug 547559 - Use Arrays.asList instead of copying array in a for loop
While there are disadvantages in some cases using this approach the
advantages are imho these:
1) It is shorter and easier to read
2) It allows to have internal optimization in future to make use of
internal parallelism (same argument as for lambda expressions) see
internal iteration
https://docs.oracle.com/javase/tutorial/collections/streams/
3) Because the size is known beforehand when calling addAll there is no
need to increase the size of the list while processing the loop Already
java 9 includes performance improvements, see
https://bugs.openjdk.java.net/browse/JDK-8155600 I am sure there is more
to come. This expression is much easier to optimize for the compiler
than a loop.

Change-Id: I53fa69661b8451d9c2d176ce192d5a6393229fbc
Signed-off-by: Carsten Hammer <carsten.hammer@t-online.de>