Bug 547720 - Update minimum SWT requirement for JFace due to API removal from Bug 475833
Requires update in SWT first.

Via Bug 475833 we removed TableTree from SWT. The current process is
that for scheduled and planned deletions we do not increase the major
version.

But the new JFace should require at least the latest SWT as it also was
adjusted to not support the TableTree widget anymore.


Change-Id: I1dc1e638682e47bd09002eb137df0b1e4a9a495c
Signed-off-by: Lars Vogel <Lars.Vogel@vogella.com>