Bug 547240: create GridDataFactory for new GridData(int style)
There is a lot of code in platform working with
new GridData(int style). There is a lot of magic in that
method with regards to the number of possibilities of creating
different GridData with just that one constructor.
Replacing these constructor calls with GridDataFactory would
end up in much longer code.
Hence GridDataFactory needs a new static method which calls
this constructor.

Change-Id: I8edf60d36dc2f781c0f4898eb128897a8d409445
Signed-off-by: Marcus Hoepfner <marcus.hoepfner@sap.com>