Bug 546932 - Key Bindings are unexpectedly moved to other BindingTables
It was incorrectly assumed that there is a one-to-one relation between
MBindingContext and MBindingTable. However, there can be multiple
binding tables for one binding context. Therefore, a random binding
table was used to update/add key bindings. As result, key bindings moved
to different binding tables.
Fix: Lookup binding tables by their elementId, instead of a table with a
reference to a binding context with that elementId.

Change-Id: I150bd0ee65a3f29df564b373616ab341700b2a95
Signed-off-by: Rolf Theunissen <rolf.theunissen@gmail.com>