Bug 548173 - Tips dialog not shown with org.eclipse.tips.startup.default
The default configuration area preference is set to "background" in
TipsPreferences.initializeDefaultPreferences(). So when checking whether
a dialog is needed, DefaultTipManager.mustOpen() will return "no",
regardless of specifying -Dorg.eclipse.tips.startup=dialog VM property
when launching Eclipse.

This change ensures that when the default start-up behavior is written
to the configuration area, the VM proprety is also taken into account.

Change-Id: I667f7fad0bb8491f48cee627c5af1dbd9a9ecd3d
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>