Bug 417255 - Race Condition in DecorationScheduler
Decoration requests that were scheduled in quick succession could result
in requests getting lost, because scheduling attempts were ignored if
the previous decoration request had not yet been updated to the UI.

Test case added in this commit and Bugzilla document the scenarios in
more detail.

To solve this, the collection of decoration results was made thread-safe
and some detection logic was added for whether it is the update job or
other logic checking that a decoration is ready before scheduling the
request. Also adding the same decoration context to an existing
decoration request multiple times is prevented to simplify result
caching.

Change-Id: Ida26e8856ec5fed5cca3f5d71b2327c6b4750166
Signed-off-by: Stefan Winkler <stefan@winklerweb.net>