Bug 548544 - Enlarge height of progress area
With bigger font sizes the progress area of the splash screen can cut
off the label. Enlarge the height of the progress area by 5 pixels, such
that the progress message may "touch" the progress bar.

The positioning of text and progress bar does not change. See bug for
screenshot with changes.

Change-Id: Ica78230a4aeae4e7cd5cd1b695f37eff9f2e3115
Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>