Bug 547754 - Slow loading "org.eclipse.e4.ui.swt.css.theme" when using shared drive as workspace directory
Currently we go through all the bundles for Css Theme contributions.
Instead it would be better if we get the contributions to the Extention
point "org.eclipse.e4.ui.css.swt.theme" and get the plugins that these
contributions belong to and only search in these plugins for the theme
related information. This patch implements the above idea. Also added
code to consider "org.eclipse.ui.themes" extension contributions.

Tested this patch for Eclipse 4.4 and also updated the plugin versions

Change-Id: I491592cce79eccee96b2b643fe287d8fb9e54139