Bug 547519 - [code mining] Don't apply on tab
This is more a workaround as the solution is not optimal at usage (caret
not properly placed), but the proposed state is still better than
current one.

This prevents from making a '\t' character "host" the annotation as it
comes with a bunch of issues in the DrawingStrategy (that fails at
properly computing the width of a tab) and in SWT (bug 547532).

Change-Id: I897da06cd57aef80a862ba28dc75887c212d234a
Signed-off-by: Mickael Istria <mistria@redhat.com>