Bug 433914 - Tabbed Properties View (TabbedPropertyList) throws NPE
This eliminates the single case where elements is null and hopefully fix
the reported NPE.
It was only null on initialization and before elements were set the
first time. There was no way (not even removing all elements) to made
the array null again and basically the entire class is based on
elements.length without null checks.

Change-Id: Ib3088fe6a033d1c1c1ef4bea3096281f0d89c687
Signed-off-by: Paul Pazderski <paul-eclipse@ppazderski.de>