Bug 549693 - Workbench: Stack information is lost in error logging / Get rid of StatusUtil.getCause()
When logging, using the cause of an Exception instead of the Exception
itself is not good practice, because valuable stack information gets
lost.

This mostly affects
    StatusUtil.newStatus(int, String, Throwable)
    StatusUtil.newStatus(String, String, Throwable)
which is used in many places in the workbench.

Therefore, many error log entries written by workbench code miss the
outermost Exception layer, which makes problems harder to track down.

Solution:
- Always log the actual Exceptions at hand, don't blindly access their
  cause.
- Get rid of StatusUtil.getCause() as it serves no purpose anymore.

Change-Id: I7d16a05b28aed7cb512c595021591bedec0c14af