Bug 527689 - CompatibilityEditor Parts are leaking in workbench.xmi
Parts related to CompatibilityEditor are not removed from the workbench
model because REMOVE_ON_HIDE_TAG is only set when a editor becomes
active (i.e. the CompatibilityEditor is instantiated).
Fix: set REMOVE_ON_HIDE_TAG on the part descriptor such that it is
always applied.

Also cleanup any CompatibilityEditor that should have been removed from the model, when the workbench is launched.

Change-Id: I5636da6ed4263779e392ea66da78ff38fc0dd761
Signed-off-by: Rolf Theunissen <rolf.theunissen@gmail.com>