Fix layout of quicksearch preference page
The multi-line preferences for ignored names/prefixes/extensions do not
resize and therefore have a different size than the maxLineLength
preference at the top of the page.

Remove the character width limitation, and set a layout data width
instead. This makes the field editors resize as expected, such that all
field editors have the same size.

Change-Id: I8e5c7b186e7058e14949ee528f336011953a786d
Signed-off-by: Michael Keppler <michael.keppler@gmx.de>