Bug 549966 - Import wizard browses twice on directory field Enter key
When hitting enter in the Import Existing Projects wizards directory or
archive fields, the list of projects that are to be imported is
retrieved from that location. During this operation, the SWT page
control is disabled, leading to a focus lost event on the same directory
or archive field. This then triggers a second browsing of the directory
or archive.

The same problem can occur when clicking the radio buttons for directory
/ zip locations. Possibly also the respective browse buttons can cause
this.

This change ensures the focus lost event is ignored, while another event
causes the wizard page to look for projects in a location. With this,
the second traversal of the import
location is avoided.

Change-Id: If37cac3132bb7e4bac431b556e4839889c0d6c83
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>