Bug 549898 - Opening "Customize perspective" hides toolbar items
Do not try to mimic the toolbar structure in the CPD, instead use the E4
model to populate the CPD.
Make the E4 model complete by populating all legacy actionsets into the
model.

This prevents that that CPD steals model elements from the E4 model, and
any tool item can be showed and hidden now.

Change-Id: Ic4c90d2afab0cabea0766769db62f9601c897d02
Signed-off-by: Rolf Theunissen <rolf.theunissen@gmail.com>
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>