Bug 550220 Sometimes incomplete stack is shown for threads that were expanded before resume
CheckState model delta visitor removes delta if parent flags are
NO_CHANGE and there are no pending updates reported for it from content
provider. In case of this bug, the updates for elements (threads in
view) are a bit late and CheckState visitor removes that delta. The
removal of delta causes improper view like expanded thread with only one
frame.

Change-Id: Ia1d33d6e59db285dbd6c1acf4503b7406809ecdb
Signed-off-by: Umair Sair <umair_sair@hotmail.com>