Bug 550438 - MultiStringMatcher (leftmost-longest): exit earlier
We may return an existing sub-match right away on a failover if
its offset is smaller than the start offset of the path we
switched to. Store the node depth in the trie nodes; this gives
us access to this information.

New tests that verify where in the text searched the algorithm
stops searching. scan003 is the test that motivated this change.
The new code stops scanning at index 5, while the old code did
this check only on the final match on the new path at index 10.

Also slightly simplify the implementation of indexOf() by
removing the 'failover' flag.

Change-Id: I3066680450a2d5027703e06afba7833b81d12127
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>