Bug 550328 - Stabilize IEditorRegistryTest.testBug308894 to fail only on errors
Copied the idea of IOConsoleTests to record the logs and to fail if any
error happens in the log during test execution. As UI freeze monitor is
configured to write errors after 2 seconds, this will prevent the
current test failure on windows after ~ 700ms due to a performance
warning log message.

Change-Id: I8e37ecd4d6afa29a001b4d5f441c4da8c8880df4
Signed-off-by: Lars Vogel <Lars.Vogel@vogella.com>