Bug 550396 - Speed up OverviewRuler
OverviewRuler computed the full pixel height of the full text,
which is expensive if the widget doesn't have a fixed line height,
for instance when word-wrapping is enabled.

However, OverviewRuler is interested in this value only if it is
strictly smaller than the height of the editor canvas. Speed up
the computation by stopping if the summed pixel height of lines
becomes larger than the height of the canvas.

Includes of 4.14 version update

Change-Id: Iace27158689a4ffb4a345124cfa46cc4c4260f06
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>