Bug 551153 - Long running quick access providers freeze UI
Compute quick access proposal in different jobs/threads and join on
result.
The processing is still sequential at the moment. Some improvement could
be to schedule the jobs for the various providers (except the
PreviousPickProvider) in parallel to populate the result, and even to
make the result appear incrementally by triggering an update whenever a
provider completes.

Change-Id: I0208dd1ee598b0de0e660b451208bbb8d1aba9d2
Signed-off-by: Mickael Istria <mistria@redhat.com>