Bug 551320 - LineNumberRulerColumn: fix overdrawn line numbers
Some subclasses may paint outside the line range given, for instance
the LineNumberChangeRulerColumn. When painting directly into the buffer
this may lead to missing line numbers.

Always use an extra Image to paint if we're not painting the whole line
range, and copy the wanted bits. Draw directly into the buffer only if
we're re-drawing everything anyway.

Change-Id: I287711173441a12d7c393d90b20d0854e1ed5f56
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>