Bug 551341 - Remove work-around for blurry line numbers on Mac HiDPI
This work-around is no longer needed, and slows down scrolling on
Mac HiDPI quite a bit. Remove it completely, and set the minimum
required version for org.eclipse.swt to the first version in which
Cocoa bug 489451 had been fixed (3.110.100).

This reverts the changes from bug 516293.

Note that the corresponding changes in LogicalLineNumberRulerColumn in
EGit must not be undone yet. EGit needs to work on older platforms
where the work-around may still be active. It can be removed in EGit
only once the baseline of EGit becomes 4.14.

Change-Id: Ibb801c98f19f118353842cae2caad6b29852e6e9
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>