Bug 551362 - Use FussyProgressMonitor to check usage of monitor
Each resource operation has to fully use up a given monitor. Instead of
creating FussyProgressMonitor instances with getMonitor() and don't
check it further, each operation that takes a monitor gets a fresh or
reseted instance of FussyProgressMonitor, which is checked to be used
up after the invocation. For exceptional cases it sanityCheck() is used
in the catch clause.


Change-Id: Id24576781147be2f07db987ab5fef3b3c9416b7d
Signed-off-by: Karsten Thoms <karsten.thoms@itemis.de>
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>