Bug 551390 - Remove custom styling and drawing from PerspectiveSwitcher
The special drawing and styling of perspective switcher it not very
visible.

Also it is questionable why we should treat on toolbar element
special and to me the used gradient looks outdated, other users do not
see it for understand why this is handled different.

This change removes
this special handling of the perspective switcher which increases the
consistency of the Eclipse UI.

Change-Id: Iee19be9749ce4be17a09f09613901a397f03e462
Signed-off-by: Lars Vogel <Lars.Vogel@vogella.com>