Bug 448472 - Cycle while computing value LegacyHandlerService
ValueComputation.get() can be accessed from multiple threads, but has no
protection against multi-threaded access. One example of this is looking
up the handler of a command via HandlerServiceImpl.lookUpHandler(), from
a job. E.g. when closing a project, the CloseUnrelatedProjectsAction
will react to a resource change event, potentially looking up a handler
simultaneously to selection change handling for other Package Explorer
actions. One of the possible results is that an exception is thrown,
that complains about cycles.

This change retains the delegated computation without synchronization,
but storing the cached value is done in a synchronized block.

Furthermore the cycle detection is improved, so that concurrent access
is not reported as a cycle. This is done by walking the (thread local)
computation stack of the current context and looking for the current
computation, but only if the current computation is already ongoing upon
method entry; i.e. only in cases of concurrent access or actual cycles.

Change-Id: I0d211c22df0c5650922e09fb93e6aa6b9a8aff62
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>