Bug 495789 - Catch errors during update for lists and sets
Handle exceptions during the update of lists and
sets in the same way as exceptions are handled for
values: Catch and set validation result.

For errors during the actual updates this is not a
change in behaviour. For errors during conversion
and misc errors at other points this is a change.

Change-Id: Icc08aa9f1a1e4e9c55ddf3f48c4d4fd12990949b
Signed-off-by: Jens Lidestrom <jens@lidestrom.se>