Bug 551324 - Allow to hide text in Find Actions tool contribution
The text for the Find Action but can now be configured based on the user
request. This follows the same pattern as the perspective switch.

The
ability to hide this toolitem has been removed, if that is required (the
other toolitems also don't have this) we could implement this later. The
correct place to implement this would IMHO be the Customize Perspective
dialog.

Change-Id: Ic5bcb9cae10af56555b23564053f9ac94afdc24a
Signed-off-by: Lars Vogel <Lars.Vogel@vogella.com>