Bug 293408 - [WorkingSets] Moving from "Working Set Contents" page to another page raises an error
Using setPageComplete(false) to always mark the page as invalid on
creation makes sense only when the page is used in a wizard before
the user has entered any details about the working set.

This assumption is false when the page is used in the properties
dialog to show the details of a pre-existing working set. The dialog
uses this invalid state to tell the user you can't navigate away,
even though it contains perfectly valid data. You have to make some
change to cause validateInput() in order to work around it.

This change makes the page do validateInput() on creation instead
to determine the true validity of the page's data. This way the
property dialog allows the user to navigate away even if they make
no changes.

Change-Id: I2b971d6a93850902f9870214d7ce8352ea90da22
Signed-off-by: Mat Booth <mat.booth@redhat.com>