Bug 552712 - Package explorer context menu slow for big selection
Avoid iterating all elements of a multi selection to calculate whether
common property pages are available for all the elements of that
selection:
* Stop iterating further selection elements when the intersection of
applicable property pages becomes empty, since that intersection cannot
grow again.
* Since almost no property pages are applicable to multi selection (but
only to single selection), check that early instead of at the very end.
Chances are high that the result collection is empty just because of
that.

In my local tests the calculation is now always finished during the very
first iteration, instead of finishing the loop over all selection
elements.

Change-Id: If514738ad60fa7baef94889f2cf84e0fa5d3721b
Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>