Bug 552691 - [Dark theme] dialogs like rename or delete have white artifacts
So avoid styling artifacts we style all elements which are direct and
second level children in a Composite.

To avoid that this breaks part styling, we do the same stying for the
part and the active part.

Looks good in a runtime IDE, will run a custom build with this change to
verify it more.

Scenarios tested in runtime IDE:

- Open Java editor -> OK
- Use Rename dialog -> OK
- Use Java project -> OK, still a known problem with the details page,
so same situation
-> Plug-in project creation wizard -> OK
-> Preference  clicked through a lot -> OK
-> Property page of project -> OK
-> HeapStatus monitor -> OK
-> Starting an RCP application and tabs stay "white" and do not grayed
out
-> Starting an runtime IDE and tabs labels stay "white" and are not
grayed out


Change-Id: Idfe50da6d550085568fa88e3b81bd530f44f5822
Signed-off-by: Lars Vogel <Lars.Vogel@vogella.com>