Bug 552905 - Shorter text for "Are you sure you want to delete ..."
The text in Eclipse IDE dialogs and EGit have been in lots of places
shorted to remove unnecessary text, which is not helpful to the user and
only takes time to read. To align with platform and EGit and to reduce
the time to read the text the delete messages in the refactoring dialogs
should also get shorter.

Change-Id: Ia3934a74ebc26ecb761506b8f8f0c73a8b65577a
Signed-off-by: Lars Vogel <Lars.Vogel@vogella.com>