Bug 552652 - [Progress] Progress bars flicker on Windows
Progress view will often dispose and recreate progress items. E.g. on
some refreshes or any time an element is added. The new progress bar is
initialized with zero progress and then set to the current job progress.
On Windows if a progress bar is set to a higher value an animation is
enforced. Therefore every time a new item is added any existing progress
bar jumps to empty and animates back to its previous value resulting in
flickering.
This is usually no problem on Linux because the jumps from current
progress to zero progress back to current progress are not animated and
therefore not visible.

Change-Id: Iebb33d20420a90c237476edaa648e01672ab8ad4
Signed-off-by: Paul Pazderski <paul-eclipse@ppazderski.de>