Bug 553133 - VisibleLinesTracker: don't mix model lines and pixels
VisibleLinesTracker called JFaceTextUtil.getLinePixel(widget, index)
with a model line index. This is wrong and caused trouble with line
numbers not updating correctly in some cases when folding/unfolding
in a ProjectionViewer.

Track both widget and model bottom indices, and use the widget index
to determine the bottom pixel.

Also update copyright year and use the standard copyright header
format.

Change-Id: Ia4ad93078cd1e278b4f496277bf00d53dda1e7e3
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>