Bug 553089 - Avoid loading / parsing of log file in UI thread
Parsing the log file is a potential long running task depending on the
log's size. It has to be avoided that this is performed on the UI
thread, since this would block the UI when the Error View is
initialized.
The loading is performed on the common fork-join pool with a
CompletableFuture. On completion the elements and fields collections can
be resetted, filled and limited. This does not affect UI, so it can be
performed again on a free executor. Finally, the viewer is refreshed.
This is performed on the UI thread with Display#asyncExec.
While the log is loaded the content description shows the message
"Loading..." now, which gets replaced with the appropriate title when
reading is complete. readLog() may be called from the Equinox Log
Thread, which makes it necessary to dispatch setting of the 'Loading'
label to the display thread.

Refactoring: Remove synchronization on fields 'elements' and 'groups' by
changing them to CopyOnWriteArrayList and ConcurrentHashMap.

Refactoring: Change return type of AbstractEntry#getChildren to the more
concrete type AbstractEntry. This leads to a refactoring of
dateComparator in LogView#limitEntriesCount.


Change-Id: I0dd85c7fe2570385e1835be0752130f51fd89a68
Signed-off-by: Karsten Thoms <karsten.thoms@itemis.de>