Bug 553354 - Catch SecurityException if setting sysprop is forbidden
A security manager might disallow setting of system properties. As there
are no clients of that rely on this property in the platform, the
combination of external clients together with a restrictive security
manager is unlikely. When clients have a restrictive security manager
there error log should not be polluted with error entries. Thus the
potential SecurityException is silently ignored.

Change-Id: I6cfa04cbed0537e7c92ad42284de56f81c77b97e
Also-by: Karsten Thoms <karsten.thoms@itemis.de>