Bug 553242 - [Dark theme] Preference pages and Wizards with Table/Tree look bad
Uses the same color for trees and tables as for the parts, which gives
them a strong contrast to the other widgets in dialogs. Will not be used
for parts, as the parts styling as higher priority.

Also adds styling for List as well to make for  example, Preferences >
Java > Editor -'Appearance color options' List show a contrasting
background.

Change-Id: I932ed269467caa07b2512c6bc17bd593fe7d4ca7
Signed-off-by: Lars Vogel <Lars.Vogel@vogella.com>