Bug 553770 - Error when closing output stream of disposed MessageConsole
The fix for bug 551745 adds logging, in case an IOConsole stream is
closed after all streams of the console are considered closed (e.g. on
dispose()).

For a MessageConsole, its possible to create a new MessageConsoleStream
and close it after the MessageConsole is disposed. This will produce the
new error log entry.

This change adds any MessageConsoleStream created by
MessageConsole.newMessageStream() to IOConsole.openStream, so that the
streams created by MessageConsole are also correctly handled by
IOConsole.

Change-Id: I09156248fac6b6f63daab88b8881a197e7fe83fb
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>