Bug 233667 - IllegalStateException in ProjectContentTypes
This exception appears every now and then in SWTBot tests. It's
basically a race condition, where one thread tries to calculate project
content types, while another thread already removes the same project
from the workspace.

As a workaround, it is sufficient to catch the exceptions thrown when
trying to access the removed node. This will lead to simply not using
project content types, therefore the exceptions don't need to be logged
or handled.

Change-Id: I589fe23e9d579c86445e5271f748b23c7edd12a5
Signed-off-by: Michael Keppler <michael.keppler@gmx.de>