Bug 558297 - do not recreate contributed extensions for every marker
Problems view queries if marker has a resolution generator for every
shown marker. This causes extensions to be created again and again, also
for every selection in the Problems view.

Similar problem is also for markers that have assotiated help providers
(not that many by default).

Speedup this by creating a cache for both types of contributions and
reuse them instead of creating them again and again.

Change-Id: I427a031813da5db71306046767d0c23b82e332d2
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>