Bug 553089 - Handle premature closing of view before load finished
When the log view is closed while reading the log, the view is disposed
and widget updates happen later. This can be simulated by a Thread sleep
in fetchLogEntries() and closing the view meanwhile.

Moved call to readLogFile() in createPartControl() down, since premature
closing calls dispose() and the controls are not created then.

Method setContentDescription overridden to assure that its logic runs on
the UI thread and uses a disposal guard.

Change-Id: Ie01c871eb24c2eb7acb7516e2907e6a5c26fe43f
Signed-off-by: Karsten Thoms <karsten.thoms@itemis.de>