Bug 558650 - Fix DelayedObservableValue stale event order
The stale event was sent before the isStale method returned true. Thus, if
isStale was called from the stalness listener it returned false. This is
confusing and cases problems for StalenessObservableValue.

This commit changes legacy behaviour, but it is unlikely to cause problems
for existing code, and it likely to avoid problems for new code.

Change-Id: Idbcae86ee1a06af70f8b4b08b10be4ea3b9b182b
Signed-off-by: Jens Lidestrom <jens@lidestrom.se>