Bug 421303 - Potential deadlock in IOConsolePartitioner.streamAppended
The deadlock requires that a separate thread is write large amounts of
content rapidly to console and while doing this holds lock(s) with UI
thread may require.

An example is that both (UI and non-UI thread) write to console using
the same IOConsoleOutputStream and therefore sharing the streams write
lock.

Change-Id: I5c72b60e765ec550928ba872f089494593475856
Signed-off-by: Paul Pazderski <paul-eclipse@ppazderski.de>