Bug 541448 - [Tests] StatusDialogManagerTest.testBug288765 fails dependent on dialog position
The test checks if details area is resized dependent on the length of
detailed message to show. But additional the dialog will limit the size
dependent on the available screen size and dialog position.
The dialog position depends on the parent shell position which (for
automated test runs) can be random dependent on operation system and
window manager.

As consequence the resulting details size for the short text can be the
same as for large text if the dialog has an unfavorable position.

This failure is more likley on smaller screen sizes which is the case
for current Windows test machine with resolution 1024x768.

Fix is to move dialog on top to test against the maximum available
space.

Change-Id: I34232fb2f050033bc0443a84d11d72fd385687c7
Signed-off-by: Paul Pazderski <paul-eclipse@ppazderski.de>