Bug 559476 - [Tests] Fail tests which leak modal shell
Leaked modal shells or dialogs can fail other tests run later. This can
be hard to debug because the failing test might not show any indications
why it fails and probably succeed if run alone without the leaking test
before.

Base class tearDown should check for such leaks, close the shell and
fail the offending test. This is not yet perfect because it only works
with tests inheriting UITestCase. If another test leak a modal shell the
next UITestCase will report the leak.

Change-Id: Ib50d0c6283f1736e7f9f8f45df4cf48c66b6a001
Signed-off-by: Paul Pazderski <paul-eclipse@ppazderski.de>